Refactor URLAndAddress to URLAndTags
URLAndAddress was tightly coupled to the Kubernetes service discovery use case;
it allowed an 'address' tag to be added to metrics to specify the k8s app in
which metrics had originated.

This commit broadens the struct so that it can contain any tag. This will clear
the way to add a container ID to identify the Mesos task metric source.