Respect network scope in prometheus plugin (#58)
* Add missing info to portlabel README

* Add a unit test for three network modes

This commit adds a test case which tests prometheus-labeled ports in all three
network modes: host, bridge and container.

Add ports and labels to test case

* Implement proper IP address retrieval

* Fix tasklabelViaIndex

* Fix tasklabelAlternatePath

* Fix portlabel

* Fix tasklabelViaName

* Fix tasklabelIndexPriority

* Remove unneeded mesosHostname field and getMesosHostname func

* Revert "Remove unneeded mesosHostname field and getMesosHostname func"

This reverts commit 1a80b57044e59d0bc823598e0054fbcc15bfcdd8.

* Correct assumptions in test cases

* Add method to retrieve metrics hostname for task

The correct hostname for a task is determined by the network-scope label on the
port. If set to 'container', the hostname is the first IP address in the last
container status. If set to 'host' or absent, the hostname is the same as the
local node.

* Call metrics hostname method in relevant methods